// src/app/shared/definitions/models/entities.model.ts
export class DesignModel {
    "id": number;
    "text": string;
    "content": string;
}

export class ColorModel {
    "id": number;
    "color": string;
    "coverage": string;
    "GSM": string;
    "l": string;
    "a": string;
    "b": string;
}

export class EmployeeModel {
    "id": number;
    "code": number;
    "fName": string;
    "lName": string;
    "department": string;
    "view": string;
    "nationalCode": string;
    "certificateId": string;
    "fatherName": string;
    "born": string;
    "birthdayDate": string;
    "gender": string;
    "maritalStatus": string;
    "childrenNumber": number;
    "bloodType": string;
    "hight": number;
    "education": string;
    "militaryStatus": string;
    "mobileNumber": string;
    "homeNumber": string;
    "emergencyPhone": string;
    "address": string;
    "insuranceCode": string;
    "bankAccountNumber": string;
    "bankName": string;
    "bankCardNumber": string;
    "email": string;
    "password": string;
    "personalPhoto": string;
    "rollCallCode": string;
    "borhanCode": string;
    "nationalCardImage": string;
    "certificationImage": string;
}